requirements.txt
certifi==2018.1.18
chardet==3.0.4
flake8==3.5.0
future==0.16.0
idna==2.6
mccabe==0.6.1
pew==1.1.2
pipenv==9.0.3
psutil==5.3.1
pycodestyle==2.3.1
pyflakes==1.6.0
pyTelegramBotAPI==3.5.2
python-telegram-bot==9.0.0
requests==2.18.4
six==1.11.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
